+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## unreleased
 
+## 0.6.0
+
 ### Added
 
-* Added progress bar to sync modal
-* Actual conflict resolution handling ([#8](https://github.com/LunarWatcher/obsidian-webdav-sync/issues/8))
+* Added a progress bar to the sync modal
+* Actual conflict resolution ([#8](https://github.com/LunarWatcher/obsidian-webdav-sync/issues/8))
 
 ## 0.5.2